add custom dns to qemu
example.com seems to have infra problems. The dns gives multiple
IP addresses, however it looks that machiens in the 96.7.128.x
range are not responding - which is a problem.

To avoid this, use a local DNS server to resolve the addresses, which
should give back a working IP for exmaple.com