update geckodriver, try to fix language test
Language test times out on non-x86 systems. The lnaguagees seem to work, but
in case all the local addon scopes are enabled, then it takes longer to start up
than usual. I guess this will need some investigation... but for now let's see
if the tests are working, so work around it.